refactor: split ProposalDetailCard into smaller components
